+++ src/ipc/api-types.ts
@@ -26,6 +26,7 @@ export enum Endpoint {
     showFolder = "showFolder",
     windowReady = "windowReady",
     getFileToOpen = "getFileToOpen",
+    getUrlToOpen = "getUrlToOpen",
     getWindowIndex = "getWindowIndex",
     openNewWindow = "openNewWindow",
     getWindowPages = "getWindowPages",
@@ -37,6 +38,10 @@ export enum Endpoint {
     getAppVersion = "getAppVersion",
     getRuntimeVersions = "getRuntimeVersions",
     setNativeTheme = "setNativeTheme",
+    registerAsDefaultBrowser = "registerAsDefaultBrowser",
+    unregisterAsDefaultBrowser = "unregisterAsDefaultBrowser",
+    isRegisteredAsDefaultBrowser = "isRegisteredAsDefaultBrowser",
+    openDefaultAppsSettings = "openDefaultAppsSettings",
 }
 
 export type Api = {
@@ -63,6 +68,7 @@ export type Api = {
     [Endpoint.showFolder]: (path: string) => Promise<void>;
     [Endpoint.windowReady]: () => Promise<void>;
     [Endpoint.getFileToOpen]: () => Promise<string | undefined>;
+    [Endpoint.getUrlToOpen]: () => Promise<string | undefined>;
     [Endpoint.getWindowIndex]: () => Promise<number>;
     [Endpoint.openNewWindow]: (filePath?: string) => Promise<number>;
     [Endpoint.getWindowPages]: () => Promise<WindowPages[]>;
@@ -74,6 +80,10 @@ export type Api = {
     [Endpoint.getAppVersion]: () => Promise<string>;
     [Endpoint.getRuntimeVersions]: () => Promise<RuntimeVersions>;
     [Endpoint.setNativeTheme]: (mode: "light" | "dark") => Promise<void>;
+    [Endpoint.registerAsDefaultBrowser]: () => Promise<void>;
+    [Endpoint.unregisterAsDefaultBrowser]: () => Promise<void>;
+    [Endpoint.isRegisteredAsDefaultBrowser]: () => Promise<boolean>;
+    [Endpoint.openDefaultAppsSettings]: () => Promise<void>;
 };
 
 export enum EventEndpoint {
@@ -87,6 +97,7 @@ export enum EventEndpoint {
     eZoomChanged = "eZoomChanged",
     eUpdateAvailable = "eUpdateAvailable",
     eOpenUrl = "eOpenUrl",
+    eOpenExternalUrl = "eOpenExternalUrl",
 }
 
 export interface EventSubscription {
@@ -109,6 +120,7 @@ export type EventApi = {
     [EventEndpoint.eZoomChanged]: EventObject<number>;
     [EventEndpoint.eUpdateAvailable]: EventObject<UpdateCheckResult>;
     [EventEndpoint.eOpenUrl]: EventObject<string>;
+    [EventEndpoint.eOpenExternalUrl]: EventObject<string>;
 };
 
 export enum RendererEvent {